RC-85 Fix delete ride post view structure so it receives and gets data sent from the front-end

+++ ridebackend/ridecars/views.py
@@ -57,12 +57,20 @@ class UpdateRidePost(APIView):
 class DeleteRidePost(APIView):
     permission_classes = [IsAuthenticated]
 
-    def delete(self,request,*,  car_post_id, image_id):
-        image_entry = get_object_or_404(File, pk=image_id)
+    def delete(self,request):
+        car_post_id = request.data.get('car_post_id')
+        image_id = request.data.get('image_id')
+
+        if car_post_id is None or image_id is None:
+            return Response({'detail': "Missing required arguments: car_post_id or image_id"})
+
+
         car_post = get_object_or_404(CarPost, pk=car_post_id)
+        image_entry = get_object_or_404(File, pk=image_id)
         if car_post.user != request.user:
             return Response({"detail": "You do not have permission to delete this car post."})
-        image_entry.delete()
+
         car_post.delete()
+        image_entry.delete()
         
         return Response()
